+++ src/org/example/AirportUpdateSubmit.java
@@ -47,7 +47,7 @@ public class AirportUpdateSubmit extends HttpServlet {
 
            double lat = Double.parseDouble(latitude);
            double lon = Double.parseDouble(longitude);
-           double fl = Double.parseDouble(fuel);
+           int fl = Integer.parseInt(fuel);
 
             ap.updateAirport(oldfaa, faa, airportName, airportType, streetAddress, city, lat, lon, contactName, contactPhone, ctaf, runwayType, towerFreq, fl, storage, airportComms);
             apnew.readFromDatabase(faa);
